requirements.txt
ccptools >= 1.1, <2

protobuf >=4.25.3, <5
grpcio >=1.62.1, <2
grpcio-tools >=1.62.1, <2
googleapis-common-protos >=1.63.0, <2
